requirements.txt
# Core
langchain==0.3.7
langchain-openai==0.2.7
langchain-community==0.3.7

# Vector DB
chromadb==0.5.23

# PubMed
biopython==1.84

# Utilities
python-dotenv==1.0.1
requests==2.32.3

# Optional: local embeddings (no OpenAI cost)
sentence-transformers==3.0.1